% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{allChildren}
\alias{allChildren}
\title{Gives children of particular elements on a nested list}
\usage{
allChildren(id, dChildren, includeSelf = F)
}
\arguments{
\item{id}{ids of the elements for which we search the children}

\item{dChildren}{object describing children}

\item{includeSelf}{should the result include the id in the children}
}
\value{
all children
}
\description{
TODO: make test for the real functioning of the function
}
