% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{appendRepListVectorElements}
\alias{appendRepListVectorElements}
\title{Add elements to a list, using mapply}
\usage{
appendRepListVectorElements(l, toAdd)
}
\arguments{
\item{l}{list}

\item{toAdd}{elements to add the the elements of l (need to be a list of the same size than l)}
}
\value{
the list with added elements
}
\description{
Add elements to a list, using mapply
}
