% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{childrenPath}
\alias{childrenPath}
\title{Gives the path corresponding to the next level of the list (from a defined path)}
\usage{
childrenPath(path, l)
}
\arguments{
\item{path}{path}

\item{l}{list}
}
\value{
list of path of levels n+1
}
\description{
Gives the path corresponding to the next level of the list (from a defined path)
}
