% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{createTableFK_statement}
\alias{createTableFK_statement}
\title{Create statement with foreign key REFERENCES clauses}
\usage{
createTableFK_statement(
  conn,
  tabName,
  fields,
  types,
  pk,
  foreignTable,
  foreignRef,
  listConstraint = list(),
  schema = NULL
)
}
\arguments{
\item{conn}{connection to a SQL database}

\item{tabName}{name of the table}

\item{fields}{name of the fields}

\item{types}{types of the fields}

\item{pk}{logical are the fields primary key (may only have one TRUE)}

\item{foreignTable}{names of the foreign table for the foreign keys}

\item{foreignRef}{names of the foreign variable for the foreign keys}

\item{listConstraint}{list, constraints to add}

\item{schema}{schema in which to create the table}
}
\value{
statement to be used to create the table
}
\description{
Create statement with foreign key REFERENCES clauses
}
