% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{exportPostgres}
\alias{exportPostgres}
\title{Export tables in postgres}
\usage{
exportPostgres(
  extractedTables,
  conn,
  schema = NULL,
  overwrite = T,
  createFKindices = T
)
}
\arguments{
\item{extractedTables}{list of tables to export}

\item{conn}{connection to the the postgres database}

\item{schema}{schema in which to create the tables}

\item{overwrite}{boolean, whether to overwrite schemas and tables}

\item{createFKindices}{Whether to create the foreign keys indexes}
}
\value{
result of the dbCommit function at the end of the transaction
}
\description{
Export tables in postgres
}
