% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{exportSQLite}
\alias{exportSQLite}
\title{Export the extracted tables in a sqlite database}
\usage{
exportSQLite(
  extractedTables,
  sqlite_file,
  overwrite = T,
  saveBAK = NULL,
  createFKindices = T
)
}
\arguments{
\item{extractedTables}{extracted tables}

\item{sqlite_file}{path and name of the sqlite file}

\item{overwrite}{whether to overwrite the sqlite database}

\item{saveBAK}{whether to create a backup of the sqlite file}

\item{createFKindices}{whether to create indexes for the foreign keys}
}
\value{
SQLite database connection
}
\description{
Export the extracted tables in a sqlite database
}
