% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{exportXL}
\alias{exportXL}
\title{Export extracted tables from a list of xml documents}
\usage{
exportXL(extractedTables, file, exportInfoTables = T)
}
\arguments{
\item{extractedTables}{extracted tables from the function extractTable}

\item{file}{path and file name from the excel file}

\item{exportInfoTables}{whether to export the tables which describe the tables}
}
\value{
result of the openxlsx::saveWorkbook
}
\description{
Export extracted tables from a list of xml documents
}
