% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extractAdressesMultiXml}
\alias{extractAdressesMultiXml}
\title{Extract indexes of an xml extraction from bash functions}
\usage{
extractAdressesMultiXml(
  text_doc_xml,
  grep_name = "^---file:(.*)---$",
  name_part = "\\\\1",
  verbose = T
)
}
\arguments{
\item{text_doc_xml}{character vector including the xml content, and separator (see grep_name)}

\item{grep_name}{regex expected to match complete elements to be considered as separators between files, a part of the regex should be considered as the name of the elements}

\item{name_part}{part of the separator regex that should be extracted as name (see second argument of sub)}

\item{verbose}{should the function give information during execution}
}
\value{
data.frame with names, begin and end indexed in the character vector containing the xml content
}
\description{
Extract indexes of an xml extraction from bash functions
}
