% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extractStructureListDocuments}
\alias{extractStructureListDocuments}
\title{Analyse the structure of a nested list resulting from the transformation of an xml structure}
\usage{
extractStructureListDocuments(
  listDocument,
  goFurther = c("list", "XMLAttributes")
)
}
\arguments{
\item{listDocument}{list of xml documents, transformed as a nested list}

\item{goFurther}{for which kind of variables should we go further in the nested levels (by default lists and xml attributes flags)}
}
\value{
an object describing the structure of a nested list
}
\description{
Analyse the structure of a nested list resulting from the transformation of an xml structure
}
