% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extractTables}
\alias{extractTables}
\title{Extract tables of variables from an analysed nested list (from a list of xml documents)}
\usage{
extractTables(
  dat,
  struct,
  gpsAndVar,
  convertMode = T,
  noConvert = c("emlVersion", "replacedEmlVersion", "version")
)
}
\arguments{
\item{dat}{nested list (transformation of a list of xml document)}

\item{struct}{analyses of the structure of dat (result of extractStructureListDocument)}

\item{gpsAndVar}{analyses of variables and groups (result of groupsAndVariables)}

\item{convertMode}{TODO:document}

\item{noConvert}{TODO: document}
}
\value{
list of extracted tables
}
\description{
Extract tables of variables from an analysed nested list (from a list of xml documents)
}
