% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{groupsAndVariables}
\alias{groupsAndVariables}
\title{Analyses the structure and contents of a nested list in terms of variables and groups of variables}
\usage{
groupsAndVariables(
  extract_struct,
  uniqueGpNames = T,
  gp0 = "xml_doc",
  simplify_var_id = T
)
}
\arguments{
\item{extract_struct}{structure extracted from a nested list (from function extractStructureListDocument)}

\item{uniqueGpNames}{names of the groups}

\item{gp0}{names of the root}

\item{simplify_var_id}{TODO: document}
}
\value{
extracted variables and groups from a nested list
}
\description{
Analyses the structure and contents of a nested list in terms of variables and groups of variables
}
