% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipt.R
\name{ipt_extractXml}
\alias{ipt_extractXml}
\title{Extract xml file contents from the datadir of an ipt instance}
\usage{
ipt_extractXml(iptFileTable, modeExtract = c("EML", "Structure", "both"))
}
\arguments{
\item{iptFileTable}{data.frame such as those returned by ipt_listFiles}

\item{modeExtract}{one of "EML" (for extracting basic Ecological Metadata Language variable from the eml.xml files),  "Structure" (for extracting structural metadata from the "resource.xml" files)}
}
\value{
a list (names as dataFolder from the IPT instance) with multiline character vector of length 1
}
\description{
Extract xml file contents from the datadir of an ipt instance
}
