% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{listWord2finalNames}
\alias{listWord2finalNames}
\title{Find names for variables, given a list of word describing the words which describe the variable in a nested structure}
\usage{
listWord2finalNames(
  lW,
  gp,
  max_nbW = max(sapply(lW, length)),
  reservedNames = c("string", "name", "int", "text", "class", "para")
)
}
\arguments{
\item{lW}{list of words}

\item{gp}{groups of variables in which to have unique names}

\item{max_nbW}{maximum number of words to use to get the names of the variables}

\item{reservedNames}{names which are invalid due to use in some languages}
}
\value{
final names of the variables
}
\description{
Find names for variables, given a list of word describing the words which describe the variable in a nested structure
}
