% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{metaListFromXml}
\alias{metaListFromXml}
\title{transform a xml document list to a nested list}
\usage{
metaListFromXml(xmlList)
}
\arguments{
\item{xmlList}{list of xml documents}
}
\value{
nested List copying the structure of the xml documents
}
\description{
transform a xml document list to a nested list
}
