% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{navMetaList}
\alias{navMetaList}
\alias{nextLevelLength}
\alias{nextLevelNames}
\title{Navigate a list of list}
\usage{
navMetaList(metaList, numpath)

nextLevelLength(metaList, numpath)

nextLevelNames(metaList, numpath)
}
\arguments{
\item{metaList}{List of list (as those exported by the XML::xmlToList function)}

\item{numpath}{numeric vector containing, for each level the index of the element to navigate}
}
\value{
gives the list or element of list corresponding to the successive indexes given in numpath
}
\description{
Navigate a list of list
}
