% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{numRep}
\alias{numRep}
\title{Gives the repeated list depending on number of elements}
\usage{
numRep(x)
}
\arguments{
\item{x}{TODO: document}
}
\value{
TODO: document
}
\description{
TODO: tests of the function, it is difficult to describe the effects of the funcion by only reading the code
}
