% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{plotGroupsAndVariables}
\alias{plotGroupsAndVariables}
\title{Plot the hierarchy of variables from an analysed objects (nested list extracted from an xml structure)}
\usage{
plotGroupsAndVariables(gpsAndVar, ...)
}
\arguments{
\item{gpsAndVar}{groups and variables resulting from the groupsAndVariable function}

\item{...}{additional arguments to pass to the plot.igraph function}
}
\value{
result of plot.igraph
}
\description{
Plot the hierarchy of variables from an analysed objects (nested list extracted from an xml structure)
}
