% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{save_in_excel}
\alias{save_in_excel}
\title{Export a list of tables in an Excel spreadsheet (xlsx file)}
\usage{
save_in_excel(file, lVar)
}
\arguments{
\item{file}{path and name of a file}

\item{lVar}{list of variables (or vector variable names)}
}
\value{
result of the openxlsx::saveWorkbook function
}
\description{
Export a list of tables in an Excel spreadsheet (xlsx file)
}
