% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{sqlize_extractedTables}
\alias{sqlize_extractedTables}
\title{From extracted R data.frame to SQL tables}
\usage{
sqlize_extractedTables(extractedTables)
}
\arguments{
\item{extractedTables}{extracted tables as resulting from the extractTable function}
}
\value{
transformed tables to work with SQL databases
}
\description{
From extracted R data.frame to SQL tables
}
