% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{sqlizeNames}
\alias{sqlizeNames}
\title{Transform a name to obtain a more classical name for a SQL format}
\usage{
sqlizeNames(x, maxBytes = 55)
}
\arguments{
\item{x}{name(s)}

\item{maxBytes}{maximum bytes for a SQL name (by default 55, for it to work with postgres restrictions)}
}
\value{
transformed names
}
\description{
Transform a name to obtain a more classical name for a SQL format
}
